(#22) renamed files to kebab-case
