+++ packages/sakuli-legacy/src/command/create-commands/create-project-command.class.ts
@@ -1,5 +1,5 @@
 import {Argv} from "yargs";
-import {createTestsuite} from "../../init-command/createStructure";
+import {createTestsuite} from "../../init-command/create-structure";
 
 export = {
     command: 'project [directory] [suiteName]',
